utils/cq/requirements.txt
# Packages needed to pylint code.  Most of these are needed for test rather than build.
# 94.0 is reporting a couple of errors so stick on 93.
avocado-framework<94
avocado-framework-plugin-result-html<94
avocado-framework-plugin-varianter-yaml-to-mux<94
clustershell
jira
paramiko
pyenchant
pylint
